Refactor allCountries into separate module
